tapgarden+tapdb: misc review fixes, plus lint
Address golangci-lint findings (lll, gofmt, whitespace, exhaustive,
gosec, govet copylocks) and two reviewer comments: use uint for
mintpublish's batchSize so the negative-value case is ruled out at
the type boundary, and use the already-computed batchKey in the
Cultivator.Cancel default arm instead of re-deriving it from the
live batch.

The copylocks fixes switch the augmenter staging-batch construction
from `*batch` to `batch.Copy()` so the atomic state field isn't
copied by value. The exhaustive cases get explicit default arms on
the BatchState switches in checkSingletonInvariant and RunRepairTool.
Rename Migration62BackfillSupplyUpdateEventKeys to
Migration62BackfillEventKeys so the aligned map entry fits the
80-column limit.

Make TestMigration62BackfillSupplyUpdateEventKeys and
TestMigration62BackfillDedupesLegacyDuplicates portable to Postgres:
the first one used the SQLite-only `?` placeholder, the second used
MAX(bytea) which isn't defined in Postgres. Switch to sqlc's
InsertSupplyUpdateEvent and split COUNT + key lookup into two
queries. Regenerate sqlc bindings whose source-comment references
the event_key column's migration number.

Add a [repair] section to sample-tapd.conf with an example for
repair.cancel-duplicate-batches; the sample-conf check requires every
tapd flag to have a default or example in the file.

+++ universe/mintpublish/publisher.go
@@ -20,12 +20,13 @@ import (
 // BatchRegistrar.
 type Publisher struct {
 	reg       universe.BatchRegistrar
-	batchSize int
+	batchSize uint
 }
 
 // NewPublisher constructs a Publisher that ships items to reg. batchSize
-// controls the number of items per UpsertProofLeafBatch call.
-func NewPublisher(reg universe.BatchRegistrar, batchSize int) *Publisher {
+// controls the number of items per UpsertProofLeafBatch call and must be
+// non-zero; the type rules out negative values at the boundary.
+func NewPublisher(reg universe.BatchRegistrar, batchSize uint) *Publisher {
 	return &Publisher{
 		reg:       reg,
 		batchSize: batchSize,
@@ -58,16 +59,16 @@ func (p *Publisher) PublishMintBatch(ctx context.Context,
 		items = append(items, item)
 	}
 
-	numTotal := len(items)
-	var sent int
-	for start := 0; start < numTotal; start += p.batchSize {
+	numTotal := uint(len(items))
+	var sent uint
+	for start := uint(0); start < numTotal; start += p.batchSize {
 		end := start + p.batchSize
 		if end > numTotal {
 			end = numTotal
 		}
 
 		chunk := items[start:end]
-		sent += len(chunk)
+		sent += uint(len(chunk))
 
 		log.Infof("Inserting %d new leaves (%d of %d) into local "+
 			"universe", len(chunk), sent, numTotal)
